Improved dashboard page
